+++ src/ServiceBus/Helpers/ServiceBusPurger.cs
@@ -20,7 +20,6 @@
 #endregion
 
 using Azure.Messaging.ServiceBus;
-using Azure.Messaging.ServiceBus.Administration;
 
 using System;
 using System.Collections.Generic;
@@ -256,132 +255,6 @@ namespace ServiceBusExplorer.ServiceBus.Helpers
         protected abstract ServiceBusReceiver CreateServiceBusReceiver(TEntity entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead);
 
         protected abstract Task<ServiceBusSessionReceiver> CreateServiceBusSessionReceiver(TEntity entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead);
-    }
-
-    public class TopicSubscriptionServiceBusPurger : ServiceBusPurger<SubscriptionProperties>
-    {
-        public TopicSubscriptionServiceBusPurger(ServiceBusHelper2 serviceBusHelper) : base(serviceBusHelper) { }
-
-        protected override ServiceBusReceiver CreateServiceBusReceiver(SubscriptionProperties entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead)
-        {
-            return client.CreateReceiver(
-                entity.TopicName,
-                entity.SubscriptionName,
-                new ServiceBusReceiverOptions
-                {
-                    PrefetchCount = 50,
-                    ReceiveMode = ServiceBusReceiveMode.ReceiveAndDelete,
-                    SubQueue = purgeDeadLetterQueueInstead ? SubQueue.DeadLetter : SubQueue.None
-                });
-        }
-
-        protected async override Task<ServiceBusSessionReceiver> CreateServiceBusSessionReceiver(SubscriptionProperties entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead)
-        {
-
-            return await client.AcceptNextSessionAsync(
-                entity.TopicName,
-                entity.SubscriptionName,
-                new ServiceBusSessionReceiverOptions
-                {
-                    PrefetchCount = 10,
-                    ReceiveMode = ServiceBusReceiveMode.ReceiveAndDelete
-                })
-                .ConfigureAwait(false);
-        }
-
-        protected override bool EntityRequiresSession(SubscriptionProperties entity)
-        {
-            return entity.RequiresSession;
-        }
-
-        protected override string GetEntityPath(SubscriptionProperties entity)
-        {
-            return $"{entity.TopicName}/subscriptions/{entity.SubscriptionName}";
-        }
-
-        protected async override Task<long> GetMessageCount(SubscriptionProperties entity, bool deadLetterQueueData)
-        {
-            var client = new ServiceBusAdministrationClient(serviceBusHelper.ConnectionString);
-
-            if (deadLetterQueueData)
-            {
-
-                var runtimeInfoResponse = await client.GetSubscriptionRuntimePropertiesAsync(
-                    entity.TopicName,
-                    entity.SubscriptionName)
-                    .ConfigureAwait(false);
-
-                return runtimeInfoResponse.Value.DeadLetterMessageCount;
-            }
-            else
-            {
-                var runtimeInfo = await client.GetSubscriptionRuntimePropertiesAsync(
-                    entity.TopicName,
-                    entity.SubscriptionName)
-                    .ConfigureAwait(false);
-
-                return runtimeInfo.Value.ActiveMessageCount;
-            }
-        }
-    }
-
-    public class QueueServiceBusPurger : ServiceBusPurger<QueueProperties>
-    {
-        public QueueServiceBusPurger(ServiceBusHelper2 serviceBusHelper) : base(serviceBusHelper) { }
-
-        protected override ServiceBusReceiver CreateServiceBusReceiver(QueueProperties entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead)
-        {
-            return client.CreateReceiver(
-                entity.Name,
-                new ServiceBusReceiverOptions
-                {
-                    PrefetchCount = 50,
-                    ReceiveMode = ServiceBusReceiveMode.ReceiveAndDelete,
-                    SubQueue = purgeDeadLetterQueueInstead ? SubQueue.DeadLetter : SubQueue.None
-                });
-        }
-
-        protected async override Task<ServiceBusSessionReceiver> CreateServiceBusSessionReceiver(QueueProperties entity, ServiceBusClient client, bool purgeDeadLetterQueueInstead)
-        {
-            return await client.AcceptNextSessionAsync(
-                entity.Name,
-                new ServiceBusSessionReceiverOptions
-                {
-                    PrefetchCount = 10,
-                    ReceiveMode = ServiceBusReceiveMode.ReceiveAndDelete
-                })
-                .ConfigureAwait(false);
-        }
-
-        protected override bool EntityRequiresSession(QueueProperties entity)
-        {
-            return entity.RequiresSession;
-        }
-
-        protected override string GetEntityPath(QueueProperties entity)
-        {
-            return entity.Name;
-        }
-
-        protected async override Task<long> GetMessageCount(QueueProperties entity, bool deadLetterQueueData)
-        {
-            var client = new ServiceBusAdministrationClient(serviceBusHelper.ConnectionString);
-
-            if (deadLetterQueueData)
-            {
-                var runtimeInfoResponse = await client.GetQueueRuntimePropertiesAsync(entity.Name)
-                    .ConfigureAwait(false);
-
-                return runtimeInfoResponse.Value.DeadLetterMessageCount;
-            }
-            else
-            {
-                var runtimeInfo = await client.GetQueueRuntimePropertiesAsync(entity.Name)
-                    .ConfigureAwait(false);
-
-                return runtimeInfo.Value.ActiveMessageCount;
-            }
-        }
     }
 
     public class PurgeOperationCompletedEventArgs : EventArgs
